Revamp final artifact packaging
Currently, every target produces a target-specific set of artifacts: a
`.deb` and `.changes` for Debian, `.rpm` for CentOS, `.tar.gz` for generic,
etc.  Change this to always have a single `.tar` archive, containing all
the build artifacts and `build-metadata.json`, which contains the table
of contents and rich build metadata.