Remove unnecessary shared library symlinks
Shared libraries are customarily installed as library.so.<major>.<minor>
and then a library.so and a library.so.<major> symlink is created.  The
latter is normally the exported SONAME, so technically there is no
reason to keep either the fully-versioned files or the unversioned
symlink, so we remove them both here, and make the SONAME file a regular
file.  This makes packages tidier, but, more importantly, acts as a
workaround for systems that are unable to cope with symlinks in archives
(looking at you, Nomad).